requirements.txt
Flask==2.2.5
flask_sqlalchemy==3.1.1
PyYAML==6.0.1
